# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TMDB Embed Player API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MediaItem {
  id: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  voteAverage: number;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  firstAirDate?: string | null;
  /** @nullable */
  mediaType?: string | null;
  popularity: number;
}

export interface SearchResults {
  results: MediaItem[];
  totalResults: number;
  totalPages: number;
  page: number;
}

export interface Genre {
  id: number;
  name: string;
}

export interface MovieDetail {
  id: number;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  voteAverage: number;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  runtime?: number | null;
  genres: Genre[];
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  revenue?: number | null;
}

export interface TvDetail {
  id: number;
  name: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  voteAverage: number;
  /** @nullable */
  firstAirDate?: string | null;
  /** @nullable */
  numberOfSeasons?: number | null;
  /** @nullable */
  numberOfEpisodes?: number | null;
  genres: Genre[];
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface Video {
  id: string;
  key: string;
  name: string;
  site: string;
  type: string;
  official: boolean;
}

export interface VideoResults {
  results: Video[];
}

export type EmbedDataMediaType =
  (typeof EmbedDataMediaType)[keyof typeof EmbedDataMediaType];

export const EmbedDataMediaType = {
  movie: "movie",
  tv: "tv",
} as const;

export interface EmbedData {
  id: number;
  mediaType: EmbedDataMediaType;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  voteAverage: number;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  runtime?: number | null;
  genres: Genre[];
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  trailerKey?: string | null;
  /** @nullable */
  trailerName?: string | null;
  /** @nullable */
  numberOfSeasons?: number | null;
  /** @nullable */
  numberOfEpisodes?: number | null;
}

export type SearchMediaParams = {
  query: string;
  type?: SearchMediaType;
  page?: number;
};

export type SearchMediaType =
  (typeof SearchMediaType)[keyof typeof SearchMediaType];

export const SearchMediaType = {
  movie: "movie",
  tv: "tv",
  multi: "multi",
} as const;

export type GetTrendingParams = {
  type?: GetTrendingType;
  timeWindow?: GetTrendingTimeWindow;
};

export type GetTrendingType =
  (typeof GetTrendingType)[keyof typeof GetTrendingType];

export const GetTrendingType = {
  all: "all",
  movie: "movie",
  tv: "tv",
} as const;

export type GetTrendingTimeWindow =
  (typeof GetTrendingTimeWindow)[keyof typeof GetTrendingTimeWindow];

export const GetTrendingTimeWindow = {
  day: "day",
  week: "week",
} as const;