# lib/api-zod/src/generated/types/embedData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TMDB Embed Player API
 * OpenAPI spec version: 0.1.0
 */
import type { EmbedDataMediaType } from "./embedDataMediaType";
import type { Genre } from "./genre";

export interface EmbedData {
  id: number;
  mediaType: EmbedDataMediaType;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  voteAverage: number;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  runtime?: number | null;
  genres: Genre[];
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  trailerKey?: string | null;
  /** @nullable */
  trailerName?: string | null;
  /** @nullable */
  numberOfSeasons?: number | null;
  /** @nullable */
  numberOfEpisodes?: number | null;
}