# lib/api-zod/src/generated/types/mediaItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TMDB Embed Player API
 * OpenAPI spec version: 0.1.0
 */

export interface MediaItem {
  id: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  voteAverage: number;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  firstAirDate?: string | null;
  /** @nullable */
  mediaType?: string | null;
  popularity: number;
}