# lib/api-zod/src/generated/types/movieDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TMDB Embed Player API
 * OpenAPI spec version: 0.1.0
 */
import type { Genre } from "./genre";

export interface MovieDetail {
  id: number;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  voteAverage: number;
  /** @nullable */
  releaseDate?: string | null;
  /** @nullable */
  runtime?: number | null;
  genres: Genre[];
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  revenue?: number | null;
}