# lib/api-zod/src/generated/types/searchResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TMDB Embed Player API
 * OpenAPI spec version: 0.1.0
 */
import type { MediaItem } from "./mediaItem";

export interface SearchResults {
  results: MediaItem[];
  totalResults: number;
  totalPages: number;
  page: number;
}