# lib/api-zod/src/generated/types/tvDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TMDB Embed Player API
 * OpenAPI spec version: 0.1.0
 */
import type { Genre } from "./genre";

export interface TvDetail {
  id: number;
  name: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  voteAverage: number;
  /** @nullable */
  firstAirDate?: string | null;
  /** @nullable */
  numberOfSeasons?: number | null;
  /** @nullable */
  numberOfEpisodes?: number | null;
  genres: Genre[];
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  status?: string | null;
}