# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TMDB Embed Player API
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: string;
  key: string;
  name: string;
  site: string;
  type: string;
  official: boolean;
}