# lib/api-zod/src/generated/types/videoResults.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TMDB Embed Player API
 * OpenAPI spec version: 0.1.0
 */
import type { Video } from "./video";

export interface VideoResults {
  results: Video[];
}